% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{draw_venn}
\alias{draw_venn}
\title{Draw Venn Diagram of two or three character vectors}
\usage{
draw_venn(x, cat.name, f.name)
}
\arguments{
\item{x}{A list of character vectors.}

\item{cat.name}{A character vector of category names}

\item{f.name}{A string. output file name.}
}
\value{
A png file. Venn Diagram of two or three sets
}
\description{
utility function to draw Venn diagram of two or three character vectors.
User need to specify category names and file name to store png file.
}
\examples{
x1 = as.character(c(1:5))
x2 = as.character(c(3:10))
draw_venn(list(x1, x2), cat.name=c("x1", "x2"), f.name="test.venn.png")
}
\author{
Seongyong Park (2020-08-18)
}
