% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{limma_deg}
\alias{limma_deg}
\title{Differential Gene Expression Analysis by Limma}
\usage{
limma_deg(gex, label)
}
\arguments{
\item{gex}{A gene x sample data frame. rownames are gene or probe ids and colnames are sample ids.}

\item{label}{A factor. Indicates class label of samples. It should be ordered as sample ids in gex data frame.}
}
\value{
DEG data frame which contains logFC, AveExpr, t, P.Value, adj.P.Val (by Benjamin-Hochberg correction)
}
\description{
utility function to get differential expression table from limma function in limma package.
}
\examples{
gex = do.call(rbind, lapply(1:100, function(i) rnorm(100, sample(1:10), 2)))
rownames(gex) = paste0("gid", 1:100)
colnames(gex) = paste0("sid", 1:100)
label = factor(c(rep("ALL", 50), rep("AML", 50)), levels=c("ALL", "AML"))
head(limma_deg(gex, label))
}
\author{
Seongyong Park (2020-08-18)
}
