% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovl.test.R
\name{ovl.test}
\alias{ovl.test}
\title{Permutation test to assess significance of distribution overlap}
\usage{
ovl.test(
  x1,
  x2,
  method = c("hist", "distfit", "knn", "kde"),
  nperm = 1000,
  seed = TRUE
)
}
\arguments{
\item{x1}{A numeric vector.}

\item{x2}{A numeric vector.}

\item{method}{method to calculate distribution overlap}

\item{nperm}{An integer. Number of permutation performed. default is 1000.}

\item{seed}{A logical value. default is TRUE. set seed for reproducibility.}
}
\value{
test statistic (overlap) and permutation p-value
}
\description{
Assess significance of distribution overlap.
User need to specify which method to use to calculate distribution overlap.
This function test whether given two samples are drawn from the same distribution or not. i.e.,
H0: complete overlap. Two distributions are the same. E(OVL) = 1.
H1: Two distributions are different.
}
\examples{
x1 = rnorm(100, 0)
x2 = rnorm(100, 1)
ovl.test(x1, x2, method = "hist", nperm = 1000)
}
\author{
Seongyong Park (2020-08-18)
}
