% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distfit.R
\name{ovl_distfit}
\alias{ovl_distfit}
\title{Calculate distribution overlap by distribution fitting}
\usage{
ovl_distfit(x1, x2, family = c("norm", "lognorm", ...))
}
\arguments{
\item{x1}{A numeric vector.}

\item{x2}{A numeric vector.}

\item{family}{A string to specify the distribution family.}
}
\description{
Calculate distribution overlap by distribution fitting.
User need to specify which distribution family to use.
This function estimates population distribution from sample distributions.
It utilizes fitdist function in fitdistrplus R package.
}
\examples{
x1 = rnorm(100, 0)
x2 = rnorm(100, 1)
ovl_distfit(x1, x2, family="norm")
}
\author{
Seongyong Park (2020-08-18)
}
