% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{ovl_hist}
\alias{ovl_hist}
\title{Calculate distribution overlap by histogram}
\usage{
ovl_hist(x1, x2)
}
\arguments{
\item{x1}{A numeric vector.}

\item{x2}{A numeric vector.}
}
\value{
estimated overlap between two distributions (range: 0 to 1)
}
\description{
Calculate distribution overlap by histogram. Width of bin is adjusted according to the number of samples.
This function compares two estimated distributions by histogram.
It utilizes hist function in R.
}
\examples{
x1 = rnorm(100, 0)
x2 = rnorm(100, 1)
ovl_hist(x1, x2)
}
\references{
This code is adopted from GCNR paper. (\href{https://ieeexplore.ieee.org/document/8918059}{paper})
The original code was matlab but I implemented it in the R.
No code base is available. please contact to Alfonso <\email{alfonso.r.molares@ntnu.no}>
The original implementation used fixed sized bin (n=100) but I used adaptive bin size
selection method proposed in PRE. (\href{https://journals.aps.org/pre/abstract/10.1103/PhysRevE.71.066208}{paper})
}
\author{
Seongyong Park (2020-08-18)
}
