% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde.R
\name{ovl_kde}
\alias{ovl_kde}
\title{Calculate distribution overlap by KDE}
\usage{
ovl_kde(x1, x2, n = 1024)
}
\arguments{
\item{x1}{A numeric vector.}

\item{x2}{A numeric vector.}

\item{n}{An integer. Number of bins in KDE.}
}
\value{
estimated overlap between two distributions (range: 0 to 1)
}
\description{
Calculate distribution overlap by KDE (Kernel Density Estimation)
This function compares two estimated distributions by KDE method.
It utilizes base density function in R.
}
\examples{
x1 = rnorm(100, 0)
x2 = rnorm(100, 1)
ovl_kde(x1, x2, n=1024)
}
\references{
This code is adopted from overlapping package (\href{https://doi.org/10.3389/fpsyg.2019.01089}{paper})
and modified by Seongyong Park.
You can find original implementation in overlapping R package.
(\href{https://github.com/masspastore/overlapping}{github})
}
\seealso{
\code{\link[overlapping]{overlap}}
}
\author{
Seongyong Park (2020-08-18)
}
