% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovltools.R
\docType{package}
\name{ovltools}
\alias{ovltools}
\title{ovltools: A package for computating distribution overlap between two sample distributions}
\description{
The ovltools package provides four methods to compute distribution overlaps:
KNN, KDE, histogram and distribution fitting.
}
\section{ovltools functions}{


ovl_knn: calculation by KNN

ovl_kde: calculation by KDE

ovl_hist: calculation by histogram

ovl_distfit: calculation by distribution fitting
}

