% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{Rvalue}
\alias{Rvalue}
\title{Generalized version of ovl_knn (Rvalue).}
\usage{
Rvalue(ds, cl, k = 5)
}
\arguments{
\item{ds}{A numeric vector.}

\item{cl}{An integer vector designating classes.}

\item{k}{An integer. Number of k in KNN}
}
\description{
Calculate distribution overlap by KNN (K-Nearest neighbor)
This function is the original implementation of Rvalue. It works with more than 3 classes.
The behavior of function is exactly the same as ovl_knn.
}
\examples{
ds = c(rnorm(100, 0), rnorm(100, 1), rnorm(100,2))
cl = c(rep(0,100), rep(1,100), rep(2,100))
Rvalue(ds, cl, k=5)
}
\references{
This code is adopted from AmRMR.
(\href{https://www.mdpi.com/2073-431X/8/2/42}{paper}) You can find their implementation in their websites.
(\href{https://bitldku.github.io/home/sw/R-value.html}{Java},
\href{https://bitldku.github.io/home/sw/AmRMR.html}{R}) Many thanks to the
professor Sejong Oh (Dankook University, \email{sejongoh@dankook.ac.kr}) who kindly share his work.
}
\author{
Seongyong Park (2020-08-18)
}
